+++ client/src/composables/sensors/animationFrameResizeObserver.js
